+++ packages/client-node/src/connection/node_base_connection.ts
@@ -11,6 +11,7 @@ import type {
   ConnOperation,
   ConnPingResult,
   ConnQueryResult,
+  JSONHandling,
   LogWriter,
   ResponseHeaders,
 } from '@clickhouse/client-common'
@@ -72,6 +73,8 @@ export interface RequestParams {
   enable_request_compression?: boolean
   // if there are compression headers, attempt to decompress it
   try_decompress_response_stream?: boolean
+  // if the response contains an error, ignore it and return the stream as-is
+  ignore_error_response?: boolean
   parse_summary?: boolean
   query: string
 }
@@ -82,6 +85,7 @@ export abstract class NodeBaseConnection
   protected readonly defaultAuthHeader: string
   protected readonly defaultHeaders: Http.OutgoingHttpHeaders
 
+  private readonly jsonHandling: JSONHandling
   private readonly logger: LogWriter
   private readonly knownSockets = new WeakMap<net.Socket, SocketInfo>()
   private readonly idleSocketTTL: number
@@ -106,6 +110,10 @@ export abstract class NodeBaseConnection
     }
     this.logger = params.log_writer
     this.idleSocketTTL = params.keep_alive.idle_socket_ttl
+    this.jsonHandling = params.json ?? {
+      parse: JSON.parse,
+      stringify: JSON.stringify,
+    }
   }
 
   async ping(params: ConnPingParams): Promise<ConnPingResult> {
@@ -421,7 +429,7 @@ export abstract class NodeBaseConnection
     const summaryHeader = response.headers['x-clickhouse-summary']
     if (typeof summaryHeader === 'string') {
       try {
-        return JSON.parse(summaryHeader)
+        return this.jsonHandling.parse(summaryHeader)
       } catch (err) {
         this.logger.error({
           message: `${op}: failed to parse X-ClickHouse-Summary header.`,
@@ -462,6 +470,7 @@ export abstract class NodeBaseConnection
         : // there is nothing useful in the response stream for the `Command` operation,
           // and it is immediately destroyed; never decompress it
           false
+    const ignoreErrorResponse = params.ignore_error_response ?? false
     try {
       const { stream, summary, response_headers } = await this.request(
         {
@@ -474,6 +483,7 @@ export abstract class NodeBaseConnection
           enable_response_compression:
             this.params.compression.decompress_response,
           try_decompress_response_stream: tryDecompressResponseStream,
+          ignore_error_response: ignoreErrorResponse,
           headers: this.buildRequestHeaders(params),
           query: params.query,
         },
@@ -531,9 +541,13 @@ export abstract class NodeBaseConnection
         this.logResponse(op, request, params, _response, start)
         const tryDecompressResponseStream =
           params.try_decompress_response_stream ?? true
+        const ignoreErrorResponse = params.ignore_error_response ?? false
         // even if the stream decompression is disabled, we have to decompress it in case of an error
         const isFailedResponse = !isSuccessfulResponse(_response.statusCode)
-        if (tryDecompressResponseStream || isFailedResponse) {
+        if (
+          tryDecompressResponseStream ||
+          (isFailedResponse && !ignoreErrorResponse)
+        ) {
           const decompressionResult = decompressResponse(_response, this.logger)
           if (isDecompressionError(decompressionResult)) {
             const err = enhanceStackTrace(
@@ -546,7 +560,7 @@ export abstract class NodeBaseConnection
         } else {
           responseStream = _response
         }
-        if (isFailedResponse) {
+        if (isFailedResponse && !ignoreErrorResponse) {
           try {
             const errorMessage = await getAsText(responseStream)
             const err = enhanceStackTrace(
@@ -789,6 +803,7 @@ type RunExecParams = ConnBaseQueryParams & {
   op: 'Exec' | 'Command'
   values?: ConnExecParams<Stream.Readable>['values']
   decompress_response_stream?: boolean
+  ignore_error_response?: boolean
 }
 
 const PingQuery = `SELECT 'ping'`