+++ packages/client-web/src/config.ts
@@ -3,6 +3,7 @@ import type {
   ConnectionParams,
   DataFormat,
   ImplementationDetails,
+  JSONHandling,
   ResponseHeaders,
 } from '@clickhouse/client-common'
 import { WebConnection } from './connection'
@@ -32,5 +33,6 @@ export const WebImpl: ImplementationDetails<ReadableStream>['impl'] = {
     _log_error: (err: Error) => void,
     response_headers: ResponseHeaders,
   ) => new ResultSet(stream, format, query_id, response_headers)) as any,
-  values_encoder: new WebValuesEncoder(),
+  values_encoder: (jsonHandling: JSONHandling) =>
+    new WebValuesEncoder(jsonHandling),
 }